move type declarations and extensions to separate files
