+++ src/type-extensions.ts
@@ -1,0 +1,14 @@
+import type {
+  HardhatContractSizerConfig,
+  HardhatContractSizerUserConfig,
+} from './types.js';
+
+declare module 'hardhat/types/config' {
+  interface HardhatConfig {
+    contractSizer: HardhatContractSizerConfig;
+  }
+
+  interface HardhatUserConfig {
+    contractSizer?: HardhatContractSizerUserConfig;
+  }
+}
