+++ src/types.ts
@@ -1,0 +1,13 @@
+export interface HardhatContractSizerConfig {
+  alphaSort: boolean;
+  runOnCompile: boolean;
+  flat: boolean;
+  strict: boolean;
+  only: string[];
+  except: string[];
+  outputFile: string;
+  unit: 'B' | 'kB' | 'KiB';
+}
+
+export type HardhatContractSizerUserConfig =
+  Partial<HardhatContractSizerConfig>;
